Fix test

+++ src/main/java/flower/store/FlowerPack.java
@@ -4,9 +4,9 @@ import flower.store.items.Item;
 
 public class FlowerPack implements Item {
     private Flower flower;
-    private int quantity;
+    private double quantity;
 
-    public FlowerPack(Flower flower, int quantity) {
+    public FlowerPack(Flower flower, double quantity) {
         this.flower = flower;
         this.quantity = quantity;
     }
